--- JST_SMD.kicad_mod
@@ -1,56 +1,0 @@
-(footprint "JST_SMD" (version 20221018) (generator pcbnew)
-  (layer "F.Cu")
-  (attr smd)
-  (fp_text reference "REF**" (at 0 -0.5 unlocked) (layer "F.SilkS")
-      (effects (font (size 1 1) (thickness 0.1)))
-    (tstamp 7b4600f5-a47f-4ff8-842c-156f2762030f)
-  )
-  (fp_text value "JST_SMD" (at 0 1 unlocked) (layer "F.Fab")
-      (effects (font (size 1 1) (thickness 0.15)))
-    (tstamp 9b6353ab-1742-492d-b5cc-d743e68a8dab)
-  )
-  (fp_text user "${REFERENCE}" (at -0.04 -4.93 unlocked) (layer "F.Fab")
-      (effects (font (size 1 1) (thickness 0.15)))
-    (tstamp 8a50ff6c-281a-42a9-80fd-488a8218814f)
-  )
-  (fp_line (start -4.29 -3.765) (end 3.71 -3.765)
-    (stroke (width 0.2) (type solid)) (layer "F.SilkS") (tstamp 10df94af-247d-483e-ab31-b032bfebf579))
-  (fp_line (start -4.29 4.035) (end -4.29 -3.765)
-    (stroke (width 0.2) (type solid)) (layer "F.SilkS") (tstamp 9b76fa7c-0c4d-44f7-b874-3ecbe0178ef3))
-  (fp_line (start -4.29 4.035) (end -3.29 4.035)
-    (stroke (width 0.2) (type solid)) (layer "F.SilkS") (tstamp 4a2082be-2e83-4e7f-a1ca-354783d8d794))
-  (fp_line (start -3.29 4.035) (end -3.29 2.635)
-    (stroke (width 0.2) (type solid)) (layer "F.SilkS") (tstamp b7323e10-4db9-48bb-a717-8ba025258678))
-  (fp_line (start 2.71 2.635) (end -3.29 2.635)
-    (stroke (width 0.2) (type solid)) (layer "F.SilkS") (tstamp 091c4a7e-09c1-407c-8d75-a39757748130))
-  (fp_line (start 2.71 2.635) (end 2.71 4.035)
-    (stroke (width 0.2) (type solid)) (layer "F.SilkS") (tstamp 418e5489-7d6b-4c69-95cd-4ff3549df8cf))
-  (fp_line (start 2.71 4.035) (end 3.71 4.035)
-    (stroke (width 0.2) (type solid)) (layer "F.SilkS") (tstamp 8d5dc48e-146b-483e-8699-541830ba25a7))
-  (fp_line (start 3.71 -3.765) (end 3.71 4.035)
-    (stroke (width 0.2) (type solid)) (layer "F.SilkS") (tstamp 1d5864ab-cb04-4a80-aaee-ac2709a7de19))
-  (fp_line (start -4.29 -3.765) (end -4.29 4.035)
-    (stroke (width 0.12) (type solid)) (layer "Dwgs.User") (tstamp 6f4108a4-ab8f-4fba-9279-bd301a3d5d35))
-  (fp_line (start -4.29 4.035) (end -3.29 4.035)
-    (stroke (width 0.12) (type solid)) (layer "Dwgs.User") (tstamp 1de74510-0764-4524-bc81-bb566272910c))
-  (fp_line (start -3.29 2.635) (end 2.71 2.635)
-    (stroke (width 0.12) (type solid)) (layer "Dwgs.User") (tstamp d6916135-c2ef-4ded-99de-a7a94f90b615))
-  (fp_line (start -3.29 4.035) (end -3.29 2.635)
-    (stroke (width 0.12) (type solid)) (layer "Dwgs.User") (tstamp 85b66a3c-0e73-45ce-a95f-175e6ed4c372))
-  (fp_line (start 2.71 2.635) (end 2.71 4.035)
-    (stroke (width 0.12) (type solid)) (layer "Dwgs.User") (tstamp 8c6ba469-43c4-4637-9fd8-095c39ca7a96))
-  (fp_line (start 2.71 4.035) (end 3.71 4.035)
-    (stroke (width 0.12) (type solid)) (layer "Dwgs.User") (tstamp a8c73c80-4def-47ac-b86a-0efe684601f5))
-  (fp_line (start 3.71 -3.765) (end -4.29 -3.765)
-    (stroke (width 0.12) (type solid)) (layer "Dwgs.User") (tstamp 4d4b0fa4-2322-4c08-b04b-3a051dbe846b))
-  (fp_line (start 3.71 4.035) (end 3.71 -3.765)
-    (stroke (width 0.12) (type solid)) (layer "Dwgs.User") (tstamp eb1c9e45-d2a5-4187-a8c8-b6c4a37a2ddf))
-  (pad "" smd rect (at -3.59 -2.115) (size 1 3) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (tstamp 0b79b001-7821-485c-815b-a91a474744bd))
-  (pad "" smd rect (at 3.01 -2.115) (size 1 3) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (tstamp 0ffee8b6-86e9-4bfb-84d5-40d3ac59f19f))
-  (pad "1" smd rect (at 0.71 4.185) (size 1.1 2) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (tstamp ab28e8f1-fc1b-4a4c-8415-55b6c9ff555c))
-  (pad "2" smd rect (at -1.29 4.185) (size 1.1 2) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (tstamp faa44986-7b8a-402f-8a17-c7371c08e571))
-)
