--- Power_Switch.kicad_mod
@@ -1,44 +1,0 @@
-(footprint "Power_Switch" (version 20221018) (generator pcbnew)
-  (layer "F.Cu")
-  (attr smd)
-  (fp_text reference "REF**" (at 0 -0.5 unlocked) (layer "F.SilkS") hide
-      (effects (font (size 1 1) (thickness 0.1)))
-    (tstamp 1a01c6c7-bde2-4a3f-9912-0a69af539f2f)
-  )
-  (fp_text value "Power_Switch" (at 0.3 -0.4 unlocked) (layer "F.Fab") hide
-      (effects (font (size 1 1) (thickness 0.15)))
-    (tstamp e9a6c6f8-4310-4ce4-bfe1-22dbc15621d0)
-  )
-  (fp_text user "SW_power" (at 4.7 -3.65 unlocked) (layer "F.SilkS")
-      (effects (font (size 1 1) (thickness 0.15)))
-    (tstamp 22b7d4e1-f6c3-4233-ab00-7ff7d0cfd38c)
-  )
-  (fp_text user "${REFERENCE}" (at 0.12 -5.2 unlocked) (layer "F.Fab")
-      (effects (font (size 1 1) (thickness 0.15)))
-    (tstamp 8eb160f1-c24e-4acd-9bb6-dc302d0a6bec)
-  )
-  (fp_rect (start -3.06 -2) (end 3.59 1.2)
-    (stroke (width 0.12) (type solid)) (fill none) (layer "F.SilkS") (tstamp 53d99a40-6f40-49d9-9dfe-94591fd6eeaa))
-  (fp_rect (start -3.06 -2) (end 3.59 1.2)
-    (stroke (width 0.12) (type solid)) (fill none) (layer "F.Fab") (tstamp 7bf28b08-8497-474e-be59-f7ad68eb6c20))
-  (fp_rect (start -0.41 -2) (end 0.64 -4.25)
-    (stroke (width 0.1) (type solid)) (fill none) (layer "F.Fab") (tstamp b98e690a-1316-4ed2-be41-8494c46add84))
-  (pad "1" smd rect (at 2.34 1.7) (size 0.9 1.2) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (tstamp ffe7585e-fcf7-4236-a65d-7a4f2d2fcf12))
-  (pad "2" smd rect (at -0.51 1.7) (size 0.9 1.2) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (tstamp 9fea87bf-b64a-46cd-aa82-a20eba3d5d51))
-  (pad "3" smd rect (at -2.01 1.7) (size 0.9 1.2) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (tstamp f6d94808-112c-4522-9f86-9c517b79197c))
-  (pad "MH1" thru_hole circle (at -1.26 -0.4) (size 0.9 0.9) (drill 0.9) (layers "*.Cu" "*.Mask")
-    (solder_mask_margin 0.1) (tstamp 4d9fbfa0-2d52-40dd-b19b-40465e74065c))
-  (pad "MH2" thru_hole circle (at 1.74 -0.4) (size 0.9 0.9) (drill 0.9) (layers "*.Cu" "*.Mask")
-    (solder_mask_margin 0.1) (tstamp a3544a98-707d-4a2d-9648-bc92a14934a3))
-  (pad "MP1" smd rect (at -3.56 0.725) (size 0.8 0.9) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (solder_paste_margin 0.2) (tstamp d7f69653-7764-4f88-a07c-4b21e7d1c4da))
-  (pad "MP2" smd rect (at 3.99 0.725) (size 0.8 0.9) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (solder_paste_margin 0.2) (tstamp 2ab207db-8676-414c-bb86-72e0db178df8))
-  (pad "MP3" smd rect (at 4.04 -1.475) (size 0.8 0.9) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (solder_paste_margin 0.2) (tstamp ce54c0ad-f556-4d94-9aa7-886d39f68791))
-  (pad "MP4" smd rect (at -3.56 -1.475) (size 0.8 0.9) (layers "F.Cu" "F.Paste" "F.Mask")
-    (solder_mask_margin 0.1) (solder_paste_margin 0.2) (tstamp a81cdd64-b8b6-4235-b920-e8fbbfdc8ad1))
-)
