+++ Shema and PCB/JOYSTICK.kicad_mod
@@ -1,0 +1,63 @@
+(footprint "JOYSTICK" (version 20221018) (generator pcbnew)
+  (layer "F.Cu")
+  (attr through_hole)
+  (fp_text reference "HW-504" (at 0 -0.5 unlocked) (layer "F.SilkS")
+      (effects (font (size 1 1) (thickness 0.1)))
+    (tstamp 281fd4ac-97f2-4196-9317-83b47b9b858e)
+  )
+  (fp_text value "JOYSTICK" (at 0 1 unlocked) (layer "F.Fab")
+      (effects (font (size 1 1) (thickness 0.15)))
+    (tstamp 309cd1ba-fa73-491f-9562-774cc624643b)
+  )
+  (fp_text user "CUTOUT!" (at -2.77 -4.39 unlocked) (layer "Edge.Cuts")
+      (effects (font (size 1 1) (thickness 0.15)) (justify left bottom))
+    (tstamp 4e5f011b-4595-4750-abef-2b89115b8f20)
+  )
+  (fp_text user "${REFERENCE}" (at 0 2.5 unlocked) (layer "F.Fab")
+      (effects (font (size 1 1) (thickness 0.15)))
+    (tstamp 7335ce9d-69f3-406f-bf53-9d1af17efcad)
+  )
+  (fp_rect (start -13.415 -17) (end 13.415 17)
+    (stroke (width 0.1) (type default)) (fill none) (layer "F.SilkS") (tstamp 3a18eaa1-92af-4e13-976e-d2d1456174ce))
+  (fp_poly
+    (pts
+      (xy 4.5 -13.25)
+      (xy 5.6 -13.25)
+      (xy 5.05 -14.3)
+    )
+
+    (stroke (width 0.1) (type solid)) (fill solid) (layer "F.SilkS") (tstamp bc21061d-ed9a-4be7-81dd-cedad2a83820))
+  (fp_line (start -13.4 -4.5) (end -7.92 -4.5)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp e77f5a65-aa1d-4ce5-8ccb-2f7967d487ba))
+  (fp_line (start -13.4 4.5) (end -13.4 -4.5)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp 0b9efc8f-6cf2-4446-a303-7970214147ab))
+  (fp_line (start -7.92 -10) (end 9.42 -10)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp f08df430-155e-4222-9549-485e406508b5))
+  (fp_line (start -7.92 -4.5) (end -7.92 -10)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp cabf5f3e-0377-4a71-a08b-d39b0d2ad9d3))
+  (fp_line (start -7.92 4.5) (end -13.4 4.5)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp 915d31a3-0f7f-4e5e-9d58-3f0673a6f9fa))
+  (fp_line (start -7.92 11) (end -7.92 4.5)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp 7e87bed1-5778-46a6-a52b-590c707b0564))
+  (fp_line (start 9.42 -10) (end 9.42 -4.5)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp 5428feb6-1d1c-4e89-8215-bf9ec65516d4))
+  (fp_line (start 9.42 -4.5) (end 11.72 -4.5)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp 8d9e3c32-20d7-4437-9a2b-d47eb8647752))
+  (fp_line (start 9.42 4) (end 9.42 11)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp 89f08f56-5df1-42d5-9454-07fd88ad0a15))
+  (fp_line (start 9.42 11) (end -7.92 11)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp e188a454-22b7-4e36-84e0-7e466b7aa328))
+  (fp_line (start 11.72 -4.5) (end 11.72 4)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp 5a9318e4-bc92-4f86-a3f4-23eca255663b))
+  (fp_line (start 11.72 4) (end 9.42 4)
+    (stroke (width 0.05) (type default)) (layer "Edge.Cuts") (tstamp 7c648dc8-db68-48dc-b195-98be7e012d4d))
+  (pad "" np_thru_hole circle (at -10.105 13.81) (size 3.22 3.22) (drill 3.22) (layers "F&B.Cu" "*.Mask") (tstamp 6de1125d-8f06-470e-b1eb-3609f1578c2e))
+  (pad "" np_thru_hole circle (at -9.875 -12.89) (size 3.22 3.22) (drill 3.22) (layers "F&B.Cu" "*.Mask") (tstamp f91615a8-2b6b-4743-9756-0e1994a15e75))
+  (pad "" np_thru_hole circle (at 9.875 -12.89) (size 3.22 3.22) (drill 3.22) (layers "F&B.Cu" "*.Mask") (tstamp 507b899d-601b-4cb8-9a72-a91c94ba43a0))
+  (pad "" np_thru_hole circle (at 10.105 13.81) (size 3.22 3.22) (drill 3.22) (layers "F&B.Cu" "*.Mask") (tstamp 8eeb29bc-070b-4293-8610-95082e8a5cca))
+  (pad "1" thru_hole circle (at 5.08 -15.3) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp 60f7f4f3-6523-43fa-a8ed-e1a73891527c))
+  (pad "2" thru_hole circle (at 2.54 -15.3) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp e836b39e-7a92-4861-9d37-924ad2253aa1))
+  (pad "3" thru_hole circle (at 0 -15.3) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp 9020113a-d0e8-482e-9983-363e408197c7))
+  (pad "4" thru_hole circle (at -2.54 -15.3) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp d0e49419-7c28-4464-a062-09ebdbf5ba2d))
+  (pad "5" thru_hole circle (at -5.08 -15.3) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp 4e1d3689-897f-4f76-a270-aaebf7a097b6))
+)
