+++ Shema and PCB/JOYSTICK.kicad_sch
