+++ Shema and PCB/LSM6DS3.kicad_mod
@@ -1,0 +1,36 @@
+(footprint "LSM6DS3" (version 20221018) (generator pcbnew)
+  (layer "F.Cu")
+  (attr smd)
+  (fp_text reference "REF**" (at 0 -0.5 unlocked) (layer "F.SilkS")
+      (effects (font (size 1 1) (thickness 0.1)))
+    (tstamp 7e7a17a4-fa01-4b6f-a197-48f216c69f38)
+  )
+  (fp_text value "LSM6DS3" (at 0 1 unlocked) (layer "F.Fab")
+      (effects (font (size 1 1) (thickness 0.15)))
+    (tstamp 8a8b880e-291d-43c9-9b88-90cb0923903e)
+  )
+  (fp_text user "${REFERENCE}" (at 0 2.5 unlocked) (layer "F.Fab")
+      (effects (font (size 1 1) (thickness 0.15)))
+    (tstamp 2c3973b1-c5f9-4726-8ce5-2313d8238a8f)
+  )
+  (fp_rect (start -12.5 -8.675) (end 12.5 8.675)
+    (stroke (width 0.1) (type default)) (fill none) (layer "F.SilkS") (tstamp 315fda2b-c915-4afb-a7b7-b4dacdcd5ca0))
+  (fp_poly
+    (pts
+      (xy 3.41 -4.5)
+      (xy 4.36 -4.5)
+      (xy 3.86 -5.4)
+    )
+
+    (stroke (width 0.1) (type solid)) (fill solid) (layer "F.SilkS") (tstamp aaec1bd8-e0a9-4adf-beb8-7ab7bdd694c2))
+  (pad "" np_thru_hole circle (at -9.745 -6.09) (size 2.97 2.97) (drill 2.97) (layers "F&B.Cu" "*.Mask") (tstamp 9cb76f52-fa9a-41e3-9a43-e5d2f3ca15f8))
+  (pad "" np_thru_hole circle (at -9.745 6.09) (size 2.97 2.97) (drill 2.97) (layers "F&B.Cu" "*.Mask") (tstamp 453f71f3-b3df-49c4-9721-32073e687208))
+  (pad "" np_thru_hole circle (at 9.745 -6.09) (size 2.97 2.97) (drill 2.97) (layers "F&B.Cu" "*.Mask") (tstamp 42dcba22-09bd-4a03-af6a-c010fac731bc))
+  (pad "" np_thru_hole circle (at 9.745 6.09) (size 2.97 2.97) (drill 2.97) (layers "F&B.Cu" "*.Mask") (tstamp 02901fa9-722e-4727-ab58-69dde52e0145))
+  (pad "1" thru_hole circle (at 3.81 -6.4) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp 69cf148c-d4ba-455d-99c4-5332fefa69c1))
+  (pad "2" thru_hole circle (at 1.27 -6.4) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp cdaf85a9-ee6c-446b-990c-de24e705f568))
+  (pad "3" thru_hole circle (at -1.27 -6.4) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp 0ea70bde-332b-4119-aa72-001a53accbb9))
+  (pad "4" thru_hole circle (at -3.81 -6.4) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp e9c235d8-623c-4fa8-9f7b-8d9cd010e244))
+  (pad "5" thru_hole circle (at 1.34 6.4) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp da1cf177-1670-46e6-8fcf-7339aeacf1ef))
+  (pad "6" thru_hole circle (at -1.24 6.4) (size 1.7 1.7) (drill 1) (layers "*.Cu" "*.Mask") (tstamp a6c5bf3e-22dc-4a12-a9af-104d1ec462aa))
+)
