+++ Shema and PCB/Schema and PCB for game
