+++ Shema and PCB/Zavrsni.kicad_pro
