+++ Shema and PCB/~Zavrsni.kicad_sch.lck
